remove some validation method from view
