+++ DriverPortal/app/src/main/java/com/edulog/driverportal/settings/changepassword/presentation/ChangePasswordView.java
@@ -1,25 +1,12 @@
 package com.edulog.driverportal.settings.changepassword.presentation;
 
 import com.edulog.driverportal.common.base.BaseView;
+import com.edulog.driverportal.settings.changepassword.model.ValidationResult;
 
 public interface ChangePasswordView extends BaseView {
+    void showValidationResult(ValidationResult validationResult);
+
     void showChangePasswordSuccess(String message);
-
-    void showInvalidDriverId(String message);
-
-    void hideInvalidDriverId();
-
-    void showInvalidOldPassword(String message);
-
-    void hideInvalidOldPassword();
-
-    void showInvalidNewPassword(String message);
-
-    void hideInvalidNewPassword();
-
-    void showPasswordDoesNotMatch();
-
-    void hidePasswordDoesNotMatch();
 
     void enableRequestChangePassword();
